backend/requirements.txt
fastapi
uvicorn
python-dotenv
aiofiles
pytest
httpx
google-cloud-logging
